refactor: relocate UI assets
