+++ demo/html/index.html
@@ -4,8 +4,8 @@
   <meta charset="utf-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <title>Deployable UI Demo</title>
-  <link rel="stylesheet" href="/static/css/theme.css" />
-  <link rel="stylesheet" href="/static/css/style.css" />
+  <link rel="stylesheet" href="/static/ui/css/theme.css" />
+  <link rel="stylesheet" href="/static/ui/css/style.css" />
 </head>
 <body>
   <div class="app">
