+++ docs/examples/entity_manager.html
@@ -3,12 +3,12 @@
 <head>
   <meta charset="utf-8" />
   <title>Entity Manager Example</title>
-  <link rel="stylesheet" href="../../src/ui/styles/window.css" />
-  <link rel="stylesheet" href="../../src/ui/styles/components.form.css" />
-  <link rel="stylesheet" href="../../src/ui/styles/components.list.css" />
-  <link rel="stylesheet" href="../../src/ui/styles/components.modal.css" />
-  <link rel="stylesheet" href="../../src/ui/styles/components.toast.css" />
-  <link rel="stylesheet" href="../../src/ui/styles/components.spinner.css" />
+  <link rel="stylesheet" href="../../src/ui/css/window.css" />
+  <link rel="stylesheet" href="../../src/ui/css/components.form.css" />
+  <link rel="stylesheet" href="../../src/ui/css/components.list.css" />
+  <link rel="stylesheet" href="../../src/ui/css/components.modal.css" />
+  <link rel="stylesheet" href="../../src/ui/css/components.toast.css" />
+  <link rel="stylesheet" href="../../src/ui/css/components.spinner.css" />
 </head>
 <body>
 <script type="module" src="./entity_manager.js"></script>
