+++ src/ui/js/components/modal.js
