+++ src/ui/js/components/spinner.js
